refactor(dial-files): split pure path helpers from home resolver
Address review feedback on the _HomePathResolver extraction:

- Move the pure path checks (validate_relative_path, reject_absolute_path)
  to _utils.py, where the DI-free helpers live; they have no dependency on
  self/DI/the per-request cache.
- _HomePathResolver now holds only the stateful, request-scoped trio
  (resolve_home_dir / resolve_appdata_url / to_display_path) shared by the
  file tools and the path-argument transformer.
- Drop the thin alias methods on _DialFileTool; repoint all call sites at
  self._home_resolver.* (stateful) and the _utils functions (pure).
- Move the resolver/validation tests into test_home_path_resolver.py.

Refs #346
