+++ src/quickapp/dial_files_tooling/_delete_file_tool.py
@@ -6,6 +6,7 @@ from quickapp.common.base_stage_wrapper import BaseStageWrapper
 from quickapp.common.exceptions import InvalidToolCallParameterException
 from quickapp.common.tool_call_result import ToolCallResult
 from quickapp.dial_files_tooling._base_file_tool import _DialFileTool
+from quickapp.dial_files_tooling._utils import reject_absolute_path
 
 
 class _DeleteFileTool(_DialFileTool):
@@ -18,9 +19,9 @@ class _DeleteFileTool(_DialFileTool):
     ) -> ToolCallResult:
         path: str = kwargs["path"]
 
-        self._reject_absolute_path("path", "delete_file", path)
-        url = await self._resolve_appdata_url(path)
-        display_path = await self._to_display_path(url)
+        reject_absolute_path("path", "delete_file", path)
+        url = await self._home_resolver.resolve_appdata_url(path)
+        display_path = await self._home_resolver.to_display_path(url)
 
         try:
             await self._dial_file_service.delete(url)
