+++ src/quickapp/dial_files_tooling/_move_file_tool.py
@@ -6,6 +6,7 @@ from quickapp.common.base_stage_wrapper import BaseStageWrapper
 from quickapp.common.exceptions import InvalidToolCallParameterException
 from quickapp.common.tool_call_result import ToolCallResult
 from quickapp.dial_files_tooling._base_file_tool import _DialFileTool
+from quickapp.dial_files_tooling._utils import reject_absolute_path
 
 
 class _MoveFileTool(_DialFileTool):
@@ -20,12 +21,12 @@ class _MoveFileTool(_DialFileTool):
         destination: str = kwargs["destination"]
         overwrite: bool = bool(kwargs.get("overwrite", False))
 
-        self._reject_absolute_path("source", "move_file", source)
-        self._reject_absolute_path("destination", "move_file", destination)
+        reject_absolute_path("source", "move_file", source)
+        reject_absolute_path("destination", "move_file", destination)
 
-        source_url = await self._resolve_appdata_url(source)
-        dest_url = await self._resolve_appdata_url(destination)
-        dest_display = await self._to_display_path(dest_url)
+        source_url = await self._home_resolver.resolve_appdata_url(source)
+        dest_url = await self._home_resolver.resolve_appdata_url(destination)
+        dest_display = await self._home_resolver.to_display_path(dest_url)
 
         try:
             await self._dial_file_service.move(
