+++ src/quickapp/dial_files_tooling/_utils.py
@@ -4,12 +4,39 @@ These have no dependency on `self`/DI, so they live here rather than on the
 `_DialFileTool` base class (which holds the download/resolution behaviour).
 """
 
+from quickapp.common.exceptions import InvalidToolCallParameterException
 from quickapp.dial_core_services.dial_file_service import FolderEntry
 
 
 def is_root_reference(path: str) -> bool:
     """Whether the path denotes the agent home directory itself ('', '.', './', '/')."""
     return path.strip() in ("", ".", "./", "/")
+
+
+def validate_relative_path(path: str) -> None:
+    """Reject relative paths that are unsafe or malformed (leading '/', '..', etc.)."""
+    if path != path.strip():
+        raise InvalidToolCallParameterException(
+            "path", "path must not have leading/trailing whitespace"
+        )
+    if path.startswith("/"):
+        raise InvalidToolCallParameterException("path", "path must not start with '/'")
+    segments = path.split("/")
+    if ".." in segments:
+        raise InvalidToolCallParameterException("path", "path must not contain '..'")
+    # Trailing '' (caused by a trailing '/') is allowed to denote a folder URL.
+    if "" in segments[:-1]:
+        raise InvalidToolCallParameterException("path", "path must not contain empty segments")
+
+
+def reject_absolute_path(parameter_name: str, tool_name: str, value: str) -> None:
+    """Reject an absolute 'files/...' URL for tools that only accept home-relative paths."""
+    if value.startswith("files/"):
+        raise InvalidToolCallParameterException(
+            parameter_name,
+            f"{tool_name} requires a relative path under agent_home_dir; "
+            "do not pass an absolute files/... URL",
+        )
 
 
 def relative_to(url: str, folder_url: str) -> str:
