+++ src/tests/unit_tests/dial_files_tooling/test_base_file_tool.py
@@ -4,127 +4,11 @@ import pytest
 from aidial_client._exception import DialException
 
 from quickapp.common.exceptions import InvalidToolCallParameterException
-from quickapp.dial_files_tooling._read_file_lines_tool import _ReadFileLinesTool
-from quickapp.dial_files_tooling._tool_configs import READ_FILE_LINES_TOOL_CONFIG
 from tests.unit_tests.dial_files_tooling._helpers import make_config, make_service
 
-
-def _tool(agent_home_dir: str = "", appdata: str | None = "appbucket"):
-    service = make_service(appdata=appdata)
-    return _ReadFileLinesTool(
-        stage_wrapper_builder=MagicMock(),
-        tool_config=READ_FILE_LINES_TOOL_CONFIG,
-        perf_timer=MagicMock(),
-        dial_file_service=service,
-        dial_files_config=make_config(agent_home_dir),
-    )
-
-
-class TestResolveAppdataUrl:
-    @pytest.mark.asyncio
-    async def test_relative_path_joined_with_home(self):
-        tool = _tool()
-        url = await tool._resolve_appdata_url("reports/summary.md")
-        assert url == "files/appbucket/reports/summary.md"
-
-    @pytest.mark.asyncio
-    async def test_subdir_home_prepended(self):
-        tool = _tool(agent_home_dir="workspace/")
-        url = await tool._resolve_appdata_url("reports/summary.md")
-        assert url == "files/appbucket/workspace/reports/summary.md"
-
-    @pytest.mark.asyncio
-    async def test_absolute_url_passes_through(self):
-        tool = _tool()
-        url = await tool._resolve_appdata_url("files/otherbucket/uploads/notes.txt")
-        assert url == "files/otherbucket/uploads/notes.txt"
-
-    @pytest.mark.asyncio
-    async def test_newline_in_relative_path_rejected(self):
-        tool = _tool()
-        with pytest.raises(InvalidToolCallParameterException) as exc:
-            await tool._resolve_appdata_url("foo\nbar")
-        assert exc.value.parameter_name == "path"
-
-    @pytest.mark.asyncio
-    async def test_newline_in_absolute_url_rejected(self):
-        tool = _tool()
-        with pytest.raises(InvalidToolCallParameterException) as exc:
-            await tool._resolve_appdata_url("files/x/foo\nbar")
-        assert exc.value.parameter_name == "path"
-
-    @pytest.mark.asyncio
-    async def test_empty_path_rejected(self):
-        tool = _tool()
-        with pytest.raises(InvalidToolCallParameterException):
-            await tool._resolve_appdata_url("")
-
-    @pytest.mark.asyncio
-    async def test_leading_slash_rejected(self):
-        tool = _tool()
-        with pytest.raises(InvalidToolCallParameterException):
-            await tool._resolve_appdata_url("/abs.md")
-
-    @pytest.mark.asyncio
-    async def test_dotdot_segment_rejected(self):
-        tool = _tool()
-        with pytest.raises(InvalidToolCallParameterException):
-            await tool._resolve_appdata_url("../escape.md")
-
-    @pytest.mark.asyncio
-    async def test_dotdot_substring_rejected(self):
-        tool = _tool()
-        with pytest.raises(InvalidToolCallParameterException):
-            await tool._resolve_appdata_url("a/../b")
-
-    @pytest.mark.asyncio
-    async def test_empty_segment_rejected(self):
-        tool = _tool()
-        with pytest.raises(InvalidToolCallParameterException):
-            await tool._resolve_appdata_url("a//b")
-
-    @pytest.mark.asyncio
-    async def test_trailing_whitespace_rejected(self):
-        tool = _tool()
-        with pytest.raises(InvalidToolCallParameterException):
-            await tool._resolve_appdata_url("notes.md ")
-
-    @pytest.mark.asyncio
-    async def test_no_appdata_raises(self):
-        tool = _tool(appdata=None)
-        with pytest.raises(InvalidToolCallParameterException) as exc:
-            await tool._resolve_appdata_url("notes.md")
-        assert "appdata" in exc.value.message
-
-    @pytest.mark.asyncio
-    async def test_folder_trailing_slash_preserved(self):
-        tool = _tool()
-        url = await tool._resolve_appdata_url("reports/")
-        assert url == "files/appbucket/reports/"
-
-
-class TestToDisplayPath:
-    @pytest.mark.asyncio
-    async def test_home_url_becomes_relative(self):
-        tool = _tool()
-        await tool._resolve_appdata_url("x.md")
-        display = await tool._to_display_path("files/appbucket/reports/summary.md")
-        assert display == "reports/summary.md"
-
-    @pytest.mark.asyncio
-    async def test_out_of_home_url_unchanged(self):
-        tool = _tool()
-        await tool._resolve_appdata_url("x.md")
-        url = "files/otherbucket/x.md"
-        display = await tool._to_display_path(url)
-        assert display == url
-
-    @pytest.mark.asyncio
-    async def test_home_url_itself_becomes_empty(self):
-        tool = _tool()
-        await tool._resolve_appdata_url("x.md")
-        display = await tool._to_display_path("files/appbucket/")
-        assert display == ""
+# Home-resolution / display-path / path-validation logic lives in
+# `_HomePathResolver` and `_utils` and is covered by test_home_path_resolver.py.
+# This module covers tool-level behaviour built on top of them.
 
 
 class TestPermissionDeniedWrapper:
